Add charset
Firefox compalains with:

Timestamp: 2013-07-01 1:06:35 PM
Error: The character encoding of the HTML document was not declared. The document will render with garbled text in some browser configurations if the document contains characters from outside the US-ASCII range. The character encoding of the page must be declared in the document or in the transfer protocol.
Source File: http://mkws.indexdata.com/
Line: 0