mkws.handle_node_with_team() now works with both DOM nodes and jQuery objects.
Previously it handled only the former. For some retarded reason
related to the JavaScript runtime, if it was given the latter (so that
node.className was undefined), pushing on and assigning
	list = classes.split(/\s+/)
would not yield an error message: instead, this function and its
callers, up several stack level, silently return. What a crock.

This took FOREVER to find.