All "targets" settings now use pz:id= instead of pz:id~ (pattern matching)
All "targets" settings now have (incorrect, but working) trailing
slashes. Why is this? The Service-proxy has code that mangles ZURLs
for the use of the CF (adding rich-database-name parameters), and its
lax regular expression wrongly interprets the trailing ":3552" as a
port number. So it "fixes" this ZURL by appending a slash, and then
uses the fixed version of the URL as its unique ID. The result is that
even though the ZURL in the Torus record is
        lui.indexdata.com:8080/solr4/select?fq=database:3552
the ID that it uses is
        lui.indexdata.com:8080/solr4/select?fq=database:3552/
And indeed if I change the MKWS code to use this ID instead of the
correct one, it can be made to work.

These should soon be replaced by much less fragile UDB-based target
filtering.