Default pazpar2_url is now undefined. That means mkws.pazpar2_url() generates it from pieces. Add new piece pp2_path, with sensible default.
I am not super-happy about how protocol-independent links are
adjusted. It turns out that for some reason it's really important to
get the leading "http:" in front of the pazpar2_url, but NOT the
service_proxy_auth URL. As a result, we have the pazpar2_url()
function consulting document.location.protocol to generate its result
(while the sp_auth_url() function doesn't need to bother), AND we
still have the adjustment code in place but now it only fires when an
explicit pazpar_url configuration item is provided by the application.
It works, but there has to be a cleaner way.