Update default configuration to use sp-mkws.indexdata.com for Pazpar2 interactions, including Service Proxy authentication. By default we no longer authenticate as mkws/mkws, but leave it for referer, hostname or IP-based authentication to work their magic.
(For now, we always send username/password=XXX/XXX unless told to use
something different by the application. This is a workaround for bug
MKSP-125 and should be changed soon!)

See the manual for details of how to set up authentication:
http://mkws.indexdata.com/mkws-manual.html#authenticating-your-mwks-application-onto-the-library

--

Note that, while existing applications that already set the config item
service_proxy_auth will continue to honour it, those that just use the
default value (mkws/mkws) WILL NO LONGER WORK. This is a BACKWARDS
INCOMPATIBLE CHANGE.

There are two ways to fix this in an application. One is to modify the
application itself so that it sets mkws_config.sp_auth_credentials to
"username/password". The other is to add an end-user record to the
library "DEMO MKWS: the MasterKey Widget Set" which recognises the URL
of the application as a referring URL.