Fix MKWS-272 ("Do not make unnecessary SP/PP2 requests")
Only set PZ2 callback for operations that we actually use. We
determine this by checking whether relevant events have been
subscribed to. (Actually, due to the way event queueing is
implemented, we also get a positive for events that have been
published, even if not subscribed to. But (A) that's going to be very
rare, and (B) even if were to happen, it would fail safe by setting up
unneeded callbacks rather than failing to set up needed ones.)