Re-add mkws-jquery.js.
I'm not sure how this got lost. It was lying around in my tools/htdocs
directory and has some git history there, but was also listed in
.gitignore and does not exist in head (so is absent from a fresh
checkout). We need to maintain it in src (with the rest of the source)
and have the Makefile copy it into tools/htdocs.