Towards MKWS-294 (Change names of MKWS classes to be case-insensitive)
The code now runs with the new-style classes. But there is lots more to do. Running applications are as ugly as sin because the CSS styles don't yet apply; use of teams has yet to be tested; backwards compatibility is only partly implemented; many examples need testing under the compatibility rules, then upgrading