Last part of MKWS-239.
The core code's private _log() function now uses JavaScript's
arguments pseudo-array to pass any sequence of arguments,
varargs-like, into the underlying JSNLog functions.

Unfortunately, it turns out that this doesn't get us anything right
now, as those underlying functions expect a single argument -- so if
you call mkws.info("found", n, "records"), it will only emit the
string "found".

For that reason, it's not worth the effort at this point to similarly
varargsify the team-level logging, but I am leaving this one done so
that the relevant code is in place when JSNLog gets updated.