Part of ACREP-35.
From the "categories" widget, remove the unnecessary and harmful check
that authentication has been completed. Of course it generally has NOT
at the time the widget is created, resulting in an assertion-failure.
But this is unnecessary anyway, because the body of the widget's code
is in any case run as a callback when authentication is complete.