Part of ACREP-35, and surely the last part?
Establish local $ as a copy of mkws.$.

This is not necessary when running using a local copy of MKWS, when $
already has that value. But for reasons that I don't understand at
all, $ is undefined when running the exact same code using a copy of
MKWS from the production server.