Proper (non-hacky) fix for SUP-807.
team.newSearch now rightly looks in the widget for configuration, not
in the team. This means that configuration items set directly in the
widget itself (such as the "target" setting in the Wikipedia and
Reference Universe widgets) are now found, whereas previously only
settings that had been established at the team level or higher -- such
as the targetfilter categories of us.openrepo.net -- were found.