Extract a ton of repeated code
The approach here is still wrong, though, in a much more fundamantal
way. It changes the colours of existing HTML elements, which means
that as soon as new records come in and the display updates, the new
elements get drawn in the old colours (which you can fix by clicking
"High Contrast" again). Instead, we should dynamically change the CSS
classes that these elements use. But that is a different commit for a
different day.